+++ hydromt/__init__.py
@@ -13,7 +13,7 @@ import warnings
 import netCDF4  # noqa: F401
 
 # submodules
-from . import _io, data_catalog, gis, model, stats, utils
+from . import _io, data_catalog, gis, model, stats
 
 # high-level methods
 from .data_catalog import DataCatalog
