+++ tests/data_catalog/drivers/geodataset/test_vector_driver.py
@@ -16,7 +16,7 @@ from hydromt.gis import vector
 class TestGeoDatasetVectorDriver:
     def test_calls_preprocess(self, mocker: MockerFixture):
         mock_geods_open: mocker.MagicMock = mocker.patch(
-            "hydromt.data_catalog.drivers.geodataset.vector_driver.open_geodataset",
+            "hydromt.data_catalog.drivers.geodataset.vector_driver._open_geodataset",
             spec=_open_geodataset,
         )
         mock_geods_open.return_value = Dataset()
@@ -71,7 +71,7 @@ class TestGeoDatasetVectorDriver:
 
     def test_calls_open_geodataset(self, mocker: MockerFixture):
         mock_geods_open: mocker.MagicMock = mocker.patch(
-            "hydromt.data_catalog.drivers.geodataset.vector_driver.open_geodataset",
+            "hydromt.data_catalog.drivers.geodataset.vector_driver._open_geodataset",
             spec=_open_geodataset,
         )
         mock_geods_open.return_value = Dataset()
